Added a service to the Android Manifest.
Created NotificationService.java to implement that service. NotificationService sends a notification whenever the service starts processing.

Added code in MainActivity.java to create a notificationChannel and a notificationManager for the application. Also added code to start the NotificationService whenever the app runs onStop()

Added some new values to strings.xml to make customizing the notification easier.