Class List Creation
Set MainActivity to just launch ClassListActivity. MainActivity.java is a useless file ans is only used to start the application.

Created ClassListActivity.java to display and manage a ListView in the related layout, activity_class_list.xml. Launches ClassInfoActivity whenever a ListItem is clicked.

Created ClassInfoActivity.java to have somewhere for ClassListActivity to start when a list item is clicked. Currently not really used but will be expanded on later in the development process. Added back button functionality to restart the ClassListActivity. Will later need to remember which activity launched this activity to return the application to its previous state when the back button is pressed.

Added activity_class_list.xml to have a view for the ClassListActivity activity. Contains only a ListView which is set to fill its container.

Added activity_class_info.xml to have a view for the ClassInfoActivity activity. Contains only temporary TextView objects.