Popup Window Creation
Created a popup window to display the class info instead of launching another activity. Currently just goes away when pressed anywhere, will be changed to have a button to close and any other buttons for further functionality. TextViews in the popup window are temporary and will be refactored to cleanly display the class information.

Created an anim file to hold custom animations. Added 2 animations that just fade the popup window in and out respectively. Modified styles to point the popup window at the animations

Removed the intent from ClassListActivity because it is no longer used. Now a layout is drawn on top of the class list instead of launching a new activity for the layout.