Cleaning up popup window
Leaving current courses as they are until firebase in setup. It is at a point where it will satisfy issue 4 (list courses by course ID) and only needs to have some firebase pull requests put in to make it a proper list.

Cleaned up the layout of the popup window and added some strings so that the information at least looks correct. This will also satisfy issue 4 and only requires some of the same firebase pull requests to make the information proper.