Remade NotificationService as a Broadcast Receiver.
Deleted NotificationService because it is not compatible with AlarmManager. Copied important code for building notification to MainActivity for a temporary home, will later be refactored into its own service class.

Added AlarmManager support to MainActivity so that it can be set up to create notifications at a specific time in the future.

Created NotificationPublisher to act as the Broadcast Receiver for the AlarmManager.

Added classes.xml to hold some class information temporarily until firebase support is set up.

Used the information in classes.xml to begin sending notifications for the users next class. Needs to be improved so that it can check all of the users classes for the soonest time instead of just the first class in their list. Also there is an issue with running this process a few minutes before the hour ends, causing an extra hour to be added before the notification is sent. Also applies to minutes and sets of 60 seconds. Will be corrected soon, left as TODOs right now.

TextView in activity_main is overwrote with some testing information currently, But doesn't need to exist as all as this is will all be refactored into service and broadcast receiver work in the end.