Cleaning up poorly named classes.
Refactored all of the Building_Listview branches classes so that the names are not just the same thing with a number added. Also refactored the layout files to match.

Added information to the building activities so that the locations of services and hours of operation of the building are both also displayed. Resolving 2 more user stories instead of just the one that was originally being resolved with this branch.

All has now been merged to master. And Building_Listview branch has been deleted.