Added Firebase Support.
Fixed ClassListActivity.class to function of Firebase data. Maintained filtering ability by using ValueEventListeners to get the firebase data into an arrayAdapter instead of using a firebaseAdapter. Modified class_info_popup.xml to reflect the data being stored in firebase.

Added FirebaseInstanceData.class to allow for easy referencing of the database.

Added ClassObject.class and UserObject.class to define Classes and Users in the database.

Modified Login.java to initialize the values in FirebaseInstanceData.class at the beginning of the apps execution.

Added FirebaseClassBuilder.class and FirebaseUserBuilder.class to add values to the database. They are basically only examples of how to use the database. They are currently unused as they have already served their purpose and have been commented out of ClassListActivity.class. If the database needs to be rebuilt then you must first delete all the information in the database, then uncomment the intents in the onCreate of ClassListActivity.class for 1 run of the application.