Changed how classes are stored.
Changed the id used to store ClassObjects to the CRN of the class instead of its title. This makes it so that each class has a unique identifier for each section.

Also changed UserObjects to store an ArrayList of Strings for classes CRNS instead of the ClassObjects to remove bloat in the database.