Refactored Visual Aid solution.
Refactored the code in the visual aid in a new branch which is updated to the current version of master. I did this for 3 reasons. To have easier to understand code, To fix the errors with the previous commit, and to integrate it smoothly with the rest of the project so that I can make a pull request on this code and do an auto merge.

Requires refactoring to display useful images instead of the ones i had on hand from the floor plans user story.

Once refactored this code will satisfy User Story #3 and the refactor is simply changing some reference values.