Added basic Logout Function
Now able to press a logout button from the class list, building list, and main menu. The button inflates a new layout that displays that the user has logged out and gived them a button to press to confirm and return to the login screen.

The functionality meets the requirements of User Story #5.

This implementation of a logout function is only good because we dont actually store any user data currently. It also potentially leads to weird situations with app flow. It needs a complete overhaul if the user system is to be developed in any way.